requirements.txt
praw==7.7.1
python-dotenv==1.0.1
polygon-api-client==1.14.2
apscheduler==3.10.4
pandas_market_calendars==4.4.1